requirements.txt
openai==0.27.0
slack-sdk==3.12.0
slack-bolt==1.9.1
python-dotenv==0.19.2
requests==2.26.0
dateparser==1.1.8